requirements.txt
fastapi
uvicorn
pillow
pytesseract
openai
matplotlib
torchvision
matplotlib
torch
python-multipart
pandas
matplotlib
nltk